Route stream-events endpoint by partition key, not organisation
The underlying collection is partitioned by (registrationId, accreditationId),
so route the admin endpoint to match. This removes the organisation lookup and
registration scan, passing path params straight to the repository.
