+++ src/plugins/router.js
@@ -23,7 +23,7 @@ import { summaryLogUploadsReportRoutes } from '#routes/v1/organisations/registra
 import * as reportsRoutes from '#reports/routes/index.js'
 import { reportsUnsubmit } from '#reports/routes/unsubmit.js'
 import { adminMeGet } from '#routes/v1/admin/me/get.js'
-import { streamEventsGet } from '#routes/v1/admin/organisations/accreditations/stream-events/get.js'
+import { streamEventsGet } from '#routes/v1/admin/registrations/accreditations/stream-events/get.js'
 import { dlqMessagesGet } from '#routes/v1/admin/queues/dlq/messages.get.js'
 import { dlqPurgePost } from '#routes/v1/admin/queues/dlq/purge.post.js'
